add an option to preserve newlines
This means a paragraph is not delimited by empty lines anymore, but by the existing newlines in the text.
This is especially useful when using justext, as this crawler is clever enough to regroup paragraphs of html text,
and preserving this segmentation avoids having a title or a button text mixed with a following sentence.
Way better behavior !!